Adds contributors to tW and tQ

+++ tq/export_md_to_pdf.sh
@@ -42,24 +42,39 @@ echo $WORKING_DIR
 pushd "$WORKING_DIR" > /dev/null
 
 # link tools folder
-ln -sf $MY_DIR/.. ./tools
+if [[ ! -L ./tools ]]; then
+  ln -sf $MY_DIR/.. ./tools
+fi
 
 repo="${LANGUAGE}_${RESOURCE}"
 url="https://git.door43.org/Door43/${repo}/archive/${TAG}.zip"
 
-wget $url -O "./${repo}.zip"
-unzip -qo "./${repo}.zip"
+if [[ ! -d $repo ]]; then
+  wget $url -O "./${repo}.zip"
+  unzip -qo "./${repo}.zip"
+  echo "Checked out repo files:"
+  ls "${repo}"
+fi
 
-echo "Checked out repo files:"
-ls "${repo}"
+version=`js-yaml "${repo}/manifest.yaml" | jq -r '.dublin_core.version'`
+publisher=`js-yaml "${repo}/manifest.yaml" | jq -r '.dublin_core.publisher'`
+issued_date=`js-yaml "${repo}/manifest.yaml" | jq -r '.dublin_core.issued'`
+title=`js-yaml "${repo}/manifest.yaml" | jq -r '.dublin_core.title'`
+checking_level=`js-yaml "${repo}/manifest.yaml" | jq -r '.checking.checking_level'`
 
-version=`yaml2json "${repo}/manifest.yaml" | jq -r '.dublin_core.version'`
-issued_date=`yaml2json "${repo}/manifest.yaml" | jq -r '.dublin_core.issued'`
-title=`yaml2json "${repo}/manifest.yaml" | jq -r '.dublin_core.title'`
-checking_level=`yaml2json "${repo}/manifest.yaml" | jq -r '.checking.checking_level'`
+while read name; do
+  name="${name%\"}"
+  name="${name#\"}"
+  if [[ ! -z $contributors ]]; then
+    contributors+='; '
+  fi
+  contributors+=$name
+done <<< $(js-yaml "${repo}/manifest.yaml" | jq -c '.dublin_core.contributor[]')
 
 echo "Current '$repo' Resource is at: ${url}"
 echo "Current '$repo' Version is at: ${version}"
+echo "Current '$repo' Publisher is: ${publisher}"
+echo "Current '$repo' Contributors are: ${contributors}"
 
 # make sure old out files are gone
 # rm -f $OUTPUT_DIR/html/*
@@ -72,8 +87,11 @@ cp "$MY_DIR/header.html" "$OUTPUT_DIR/html"
 book_export () {
     book=$1
 
-    echo "GENERATING Tempfile: $OUTPUT_DIR/html/${book}.html"
-    python -m tools.tq.md_to_html_export -i "$WORKING_DIR/$repo" -o "$OUTPUT_DIR/html" -v "$version" -d "$issued_date" -b "$book"
+    htmlfile="$OUTPUT_DIR/html/${book}.html"
+    if [[ ! -f $htmlfile ]]; then 
+      echo "GENERATING HTML File: $htmlfile"
+      python -m tools.tq.md_to_html_export -i "$WORKING_DIR/$repo" -o "$OUTPUT_DIR/html" -v "$version" -p "$publisher" -c "$contributors" -d "$issued_date" -b "$book"
+    fi
 
     headerfile="file://$OUTPUT_DIR/html/header.html"
     coverfile="file://$OUTPUT_DIR/html/cover.html"
@@ -85,8 +103,10 @@ book_export () {
         outfile="$OUTPUT_DIR/pdf/${repo}_v${version}.pdf"
     fi
     mkdir -p "$OUTPUT_DIR/pdf"
-    echo "GENERATING $outfile"
-    wkhtmltopdf --encoding utf-8 --outline-depth 3 -O portrait -L 15 -R 15 -T 15 -B 15  --header-html "$headerfile" --header-spacing 2 --footer-center '[page]' cover "$coverfile" cover "$licensefile" toc --disable-dotted-lines --enable-external-links --xsl-style-sheet "$TEMPLATE" "$bodyfile" "$outfile"
+    if [[ ! -f $outfile ]]; then
+      echo "GENERATING $outfile"
+      wkhtmltopdf --javascript-delay 2000 --encoding utf-8 --outline-depth 3 -O portrait -L 15 -R 15 -T 15 -B 15  --header-html "$headerfile" --header-spacing 2 --footer-center '[page]' cover "$coverfile" cover "$licensefile" toc --disable-dotted-lines --enable-external-links --xsl-style-sheet "$TEMPLATE" "$bodyfile" "$outfile"
+    fi
 }
 
 if [[ -z $1 ]]; then
